+++ src/commands/helps/remove.rs
@@ -1,0 +1,13 @@
+// This help page was automatically generated from the mangen.py tool.
+pub const HELP: &str = r#"Uninstall an ip from the catalog.
+
+Usage:
+    orbit remove [options] <ip>
+
+Options:
+    <ip>        ip specification to uninstall
+    --all       remove the ip from the cache and downloads
+    --recurse   fully remove the ip and its dependencies
+
+Use 'orbit help remove' to read more about the command.
+"#;