+++ src/core/ip.rs
@@ -69,11 +69,11 @@ impl From<IpArchive> for Ip {
     fn from(value: IpArchive) -> Self {
         let (man, lock, archive) = value.decouple();
         let uuid = match lock.get_self_entry(man.get_ip().get_name()) {
+            Some(entry) => entry.get_uuid().clone(),
+            None => match lock.get(man.get_ip().get_name(), man.get_ip().get_version()) {
                 Some(entry) => entry.get_uuid().clone(),
-                None => match lock.get(man.get_ip().get_name(), man.get_ip().get_version()) {
-                    Some(entry) => entry.get_uuid().clone(),
-                    None => Uuid::new()
-                }
+                None => Uuid::new(),
+            },
         };
         Self {
             mapping: Mapping::Virtual(archive),
@@ -159,8 +159,8 @@ impl Ip {
             },
             false => match lock.get(man.get_ip().get_name(), man.get_ip().get_version()) {
                 Some(entry) => entry.get_uuid().clone(),
-                None => Uuid::new()
-            }
+                None => Uuid::new(),
+            },
         };
 
         Ok(Self {