+++ src/core/lockfile.rs
@@ -171,7 +171,9 @@ pub mod v1 {
 
         /// Returns the current working ip, denoted by not having a checksum with it.
         pub fn get_self_entry(&self, target: &PkgPart) -> Option<&LockEntry> {
-            self.ip.iter().find(|&f| f.checksum.is_none() && &f.name == target)
+            self.ip
+                .iter()
+                .find(|&f| f.checksum.is_none() && &f.name == target)
         }
 
         /// Returns the highest compatible version from the lockfile for the given `target`.